move 'o' tests to separate file
